Move the HTML block into the blocks library package
