+++ packages/block-library/src/editor.scss
@@ -9,6 +9,7 @@
 @import "./file/editor.scss";
 @import "./gallery/editor.scss";
 @import "./heading/editor.scss";
+@import "./html/editor.scss";
 @import "./image/editor.scss";
 @import "./latest-comments/editor.scss";
 @import "./latest-posts/editor.scss";
