+++ packages/block-library/src/html/editor.scss
@@ -1,0 +1,12 @@
+.wp-block-html .editor-plain-text {
+	font-family: $editor-html-font;
+	font-size: $text-editor-font-size;
+	color: $dark-gray-800;
+	padding: 0.8em 1em;
+	border: 1px solid $light-gray-500;
+	border-radius: 4px;
+
+	&:focus {
+		box-shadow: none;
+	}
+}
