+++ packages/block-library/src/html/test/index.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { name, settings } from '../';
-import { blockEditRender } from '../../../packages/block-library/src/test/helpers';
+import { blockEditRender } from '../../test/helpers';
 
 describe( 'core/html', () => {
 	test( 'block edit matches snapshot', () => {
